Add files via upload
This is an implementation of an applet which draws a rectangle on mouse click and also tells the coordinates of the point of clicking w.r.t. to the top left corner of the screen.

This shows the use case where we can pass a java class to another class and make the use of the methods in the new class:

Here we create a ClickListener Class which extends the MouseAdapter and we also pass the applet to the `
Clicklistener constructor so that we can call the graphics methods present in the applet.

Issues:
We can't call the private classes of the applet class from the ClickListener class and can call only the public methods.