Add files via upload
This contains an example on how to use named inner classes inside a class in order to use the best of separate class and implementing an interface. Here we create an inner class which is extended to a mouselistener class and due to being inside the applet the mouselistener class can access both private and public var., methods of the applet class and apply them to its own class.