Add files via upload
These files contains the code for multithreading implementation in java. Here we have a counter class which is implementing the runnable class and which  the run method, and this method  counting the counter uptill a given value, ......
To make it multi threading we have added to the main app file which has the tasklist for the executors to execute and we pass on the various counter classes to this thread pool and these are run concurrently and this show how multithreading is working in java.

This displays the various counts of the thread counter.